Begin simplified implementation with rsync instead of SFTP
NET::SFTP works, but there is too much error handling and logic required
and the implementation is relatively sluggish. Decided to begin work on
a simpler version which only supports one-way sync (host to guest) and
which builds on top of existing Virtualbox shared folders using rsync
on the guest to efficiently mirror paths from the shared folder to a
path on the instance virtual disk. This will also support excluding
folders (.git directories for eg are slow to sync) and specifying that
particular paths should be symlinked back to the share. The old SFTP
implementation has been tagged in case required in future.